Update visualisation_test.js
Currently- manage to go through 3 images for tutorial and 6 images for real test. After the real test completed, it changes the window location to thankyou.php. Current problem is to close the window and redirect to thankyou.php at the same time. This can be completed through passing around variable but has not been implemented yet.